Reworked the entire file tree
