+++ cmd/echo/echo.go
@@ -1,0 +1,9 @@
+package echo
+
+import (
+	"fmt"
+)
+
+func ExecuteEcho(text string) {
+	fmt.Println(text)
+}
