+++ cmd/exit/exit.go
@@ -1,0 +1,9 @@
+package exit
+
+import (
+	"os"
+)
+
+func ExecuteExit() {
+	os.Exit(0)
+}
